% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_best_k}
\alias{get_best_k}
\title{Get optimal number of neighbors}
\usage{
get_best_k(formula, data, k, ...)
}
\arguments{
\item{formula}{Formula}

\item{data}{Data}

\item{k}{Number of neighbors}

\item{...}{Further arguments passed to bnnSurvival}
}
\value{
Optimal k
}
\description{
Get optimal number of neighbors for bnnSurvival by cross validation
}

