% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnSurvivalEnsemble.R
\docType{methods}
\name{predict,bnnSurvivalEnsemble-method}
\alias{predict,bnnSurvivalEnsemble-method}
\title{Predict survival probabilities with bagged k-nearest neighbors survival prediction.}
\usage{
\S4method{predict}{bnnSurvivalEnsemble}(object, test_data)
}
\arguments{
\item{object}{Object of class bnnSurvivalEnsemble, created with bnnSurvival().}

\item{test_data}{Data set containing data to predict survival.}
}
\description{
Predict survival probabilities with bagged k-nearest neighbors survival prediction.
}

