% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnSurvivalResult.R
\docType{methods}
\name{predictions,bnnSurvivalResult-method}
\alias{predictions,bnnSurvivalResult-method}
\title{Get Predictions}
\usage{
\S4method{predictions}{bnnSurvivalResult}(object)
}
\arguments{
\item{object}{bnnSurvivalResult object to extract predictions from}
}
\description{
Get Predictions
}

