% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{predictions}
\alias{predictions}
\title{Get Predictions}
\usage{
predictions(object, ...)
}
\arguments{
\item{object}{Object to extract predictions from}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Get Predictions
}

