% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{predictSurvProb.bnnSurvivalEnsemble}
\alias{predictSurvProb.bnnSurvivalEnsemble}
\title{Function to extract survival probability predictions from bnnSurvivalEnsemble. Use with \code{pec} package.}
\usage{
\method{predictSurvProb}{bnnSurvivalEnsemble}(object, newdata, times, ...)
}
\arguments{
\item{object}{bnnSurvivalEnsemble object.}

\item{newdata}{Data used for prediction.}

\item{times}{Not used.}

\item{...}{Not used.}
}
\value{
survival probability predictions
}
\description{
Function to extract survival probability predictions from bnnSurvivalEnsemble. Use with \code{pec} package.
}

