% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnSurvivalEnsemble.R
\docType{methods}
\name{print,bnnSurvivalEnsemble-method}
\alias{print,bnnSurvivalEnsemble-method}
\title{Generic print method for bnnSurvivalEnsemble}
\usage{
\S4method{print}{bnnSurvivalEnsemble}(x)
}
\arguments{
\item{x}{bnnSurvivalEnsemble object to print}
}
\description{
Generic print method for bnnSurvivalEnsemble
}

