% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnSurvivalResult.R
\docType{methods}
\name{print,bnnSurvivalResult-method}
\alias{print,bnnSurvivalResult-method}
\title{Generic print method for bnnSurvivalResult}
\usage{
\S4method{print}{bnnSurvivalResult}(x)
}
\arguments{
\item{x}{bnnSurvivalResult object to print}
}
\description{
Generic print method for bnnSurvivalResult
}

