% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnSurvivalEnsemble.R
\docType{methods}
\name{show,bnnSurvivalEnsemble-method}
\alias{show,bnnSurvivalEnsemble-method}
\title{Generic show method for bnnSurvivalEnsemble}
\usage{
\S4method{show}{bnnSurvivalEnsemble}(object)
}
\arguments{
\item{object}{bnnSurvivalEnsemble object to show}
}
\description{
Generic show method for bnnSurvivalEnsemble
}

