% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnSurvivalResult.R
\docType{methods}
\name{show,bnnSurvivalResult-method}
\alias{show,bnnSurvivalResult-method}
\title{Generic show method for bnnSurvivalResult}
\usage{
\S4method{show}{bnnSurvivalResult}(object)
}
\arguments{
\item{object}{bnnSurvivalResult object to show}
}
\description{
Generic show method for bnnSurvivalResult
}

