% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnSurvivalResult.R
\docType{methods}
\name{timepoints,bnnSurvivalResult-method}
\alias{timepoints,bnnSurvivalResult-method}
\title{Get timepoints}
\usage{
\S4method{timepoints}{bnnSurvivalResult}(object)
}
\arguments{
\item{object}{bnnSurvivalResult object to extract timepoints from}
}
\description{
Get timepoints
}

