% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{timepoints}
\alias{timepoints}
\title{Get Timepoints}
\usage{
timepoints(object, ...)
}
\arguments{
\item{object}{Object to extract timepoints from}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Get Timepoints
}

